Refactor and Implement tests. Implement Scala excercises.

+++ src/main/scala/no/bekk/scala/model/Team.scala
@@ -1,3 +1,3 @@
 package no.bekk.scala.model
 
-case class Team(name: String, value: Long, players: List[Player])
+case class Team(name: String, value: Double, players: List[Player])
